refactor(prize-o-tron): rename files from index to about/home/etc
