+++ lib/about/about.ts
@@ -3,8 +3,8 @@ import {Component, View, NgFor} from 'angular2/angular2';
 @Component({
   selector: "about",
   directives: [],
-  styles: [require('!raw!autoprefixer?browsers=last 2 versions!sass!./index.scss')],
-  template: require('./index.html')
+  styles: [require('!raw!autoprefixer?browsers=last 2 versions!sass!./about.scss')],
+  template: require('./about.html')
 })
 export class About {
