+++ lib/home/home.html
@@ -1,6 +1,8 @@
 <ol class="breadcrumb">
   <li class="active">Home</li>
 </ol>
+
+<shame></shame>
 <div data-ui-view="" class="ng-scope">
   <section class="col-sm-6 ng-scope">
     <ng-form name="meetupSetup" role="form" class="ng-pristine ng-invalid ng-invalid-required">
