+++ lib/navbar/navbar.html
@@ -1,4 +1,5 @@
 <nav class="navbar navbar-dark bg-inverse">
+    <user src="user"></user>
     <a class="navbar-brand" [router-link]="['./Home']">{{ title }}</a>
     <ul class="nav navbar-nav">
         <li class="nav-item">
