+++ lib/navbar/navbar.ts
@@ -1,11 +1,12 @@
 import {Component, View} from 'angular2/angular2';
 import { RouterLink} from 'angular2/router';
+import { User } from '../user/user';
 
 @Component({
   selector: "navbar",
-  directives: [RouterLink],
-  styles: [require('!raw!autoprefixer?browsers=last 2 versions!sass!./index.scss')],
-  template: require('./index.html')
+  directives: [RouterLink, User],
+  styles: [require('!raw!autoprefixer?browsers=last 2 versions!sass!./navbar.scss')],
+  template: require('./navbar.html')
 })
 export class Navbar {
     title: string;
